requirements.txt
torch
torchvision
numpy
opencv-python
pandas
matplotlib
Pillow
scipy
argparse
albumentations
pytorch3d
cuda-python